+++ backend/functests/tests.py
@@ -1,6 +1,9 @@
 """
 Functional test for the Frontend.
 
+Running the test:
+    make docker_functest
+
 NB: This test uses copy of the real database, which is deleted after test.
 NB: We do not use `self.browser.get(self.live_server_url)` because our frontend works separately.
 """
@@ -19,18 +22,15 @@ from selenium.webdriver.support.ui import WebDriverWait
 
 class NewVisitorTest(LiveServerTestCase):
     """
-    Test the real Nuxt frontend as new user, by the following algorithm:
-    1. Open the site.
-    2. Click "Sign Up" button in the navbar, and create new account.
-    3. Login.
-    4. Click "Add" menu item in the navbar.
-    5. Enter an URL, hit Enter key.
+    Test the real Nuxt frontend as new user.
     """
 
     username = "testuser123"
     password = "testpassword"
+    # Use short YouTube video to test downloading
     bookmark_url = "https://www.youtube.com/watch?v=F161MWDg-48"
     bookmark_title = "Baby Blue"
+    bookmark_title_changed = "Baby Blue - changed"
 
     def setUp(self) -> None:
         # Backup existing DB file
@@ -51,12 +51,50 @@ class NewVisitorTest(LiveServerTestCase):
         )
         self.browser.quit()
 
-    def test_sign_up(self):
+    def click_navbar_menu_bookmarks(self):
         """
-        Create new user account.
+        Click "Bookmarks" menu item in the navbar
+        """
+        menu_item = self.browser.find_element(By.ID, value="navbar-menu-item-bookmarks")
+        menu_item.click()
+        menu_item.click()
+        time.sleep(1)
+
+    def click_bookmarks_table_first_row_edit_button(self):
+        """
+        Clicks "Edit" button for the bookmark in the first row of the table.
+        """
+        button = self.browser.find_element(
+            By.XPATH,
+            value="//table[@id='table-bookmarks']/tbody/tr[1]/td[3]/button",
+        )
+        button.click()
+        time.sleep(1)
+
+    def test_all(self):
+        """
+        Run all tests in the right order, because we sign up and log in in the beginning.
+        """
+        # Sign up and log in must be done in the first place, or other tests won't work.
+        self.do_test_signup_and_login()
+        # Then we add a bookmark...
+        self.do_test_add_bookmark()
+        # Edit bookmark
+        self.do_test_edit_bookmark()
+        # These two can be executed separately (we want to run `...delete`, though, to delete downloaded file):
+        self.do_test_download_bookmark()
+        self.do_test_delete_bookmark()
+
+    def do_test_signup_and_login(self):
+        """
+        Test basic app features:
+        - Open the site.
+        - Click "Sign Up" button in the navbar, and create new account.
+        - Login.
         """
         self.browser.get(
-            settings.FRONTEND_URL,
+            # We assume that the app is running from Docker Compose on `localhost`
+            "http://localhost/",
         )
         time.sleep(1)
 
@@ -97,8 +135,13 @@ class NewVisitorTest(LiveServerTestCase):
 
         self.assertIn("Your Bookmarks", self.browser.title)
 
+    def do_test_add_bookmark(self):
         """
         Test Add bookmark page.
+        - Click "Add" menu item in the navbar.
+        - Enter an YouTube URL, hit Enter key.
+        - Click "Bookmarks" in the navbar.
+        - Check that the bookmark is listed in the table.
         """
         # Click "Add" menu item in the navbar
         menu_item = self.browser.find_element(By.ID, value="navbar-menu-item-add")
@@ -111,32 +154,81 @@ class NewVisitorTest(LiveServerTestCase):
         inputbox.send_keys(Keys.ENTER)
         time.sleep(2)
 
-        # Click "Bookmarks" menu item in the navbar
-        menu_item = self.browser.find_element(By.ID, value="navbar-menu-item-bookmarks")
-        menu_item.click()
-        menu_item.click()
-        time.sleep(1)
+        self.click_navbar_menu_bookmarks()
 
         table = self.browser.find_element(By.ID, value="table-bookmarks")
         self.assertIn(self.bookmark_title, table.text)
 
-        # Find "Edit bookmark" button for added bookmark
-        button = self.browser.find_element(
-            By.XPATH,
-            value="//table[@id='table-bookmarks']/tbody/tr[1]/td[3]/button",
-        )
+    def do_test_edit_bookmark(self):
+        """
+        Edit the bookmark using EditBookmarkModal.
+        """
+        self.click_navbar_menu_bookmarks()
+        self.click_bookmarks_table_first_row_edit_button()
+
+        # Change bookmark's title and save it
+        inputbox = self.browser.find_element(By.ID, value="input-title")
+        inputbox.click()
+        # NB: `COMMAND` may be changed to `CONTROL` on Windows
+        inputbox.send_keys(Keys.COMMAND, "a")
+        inputbox.send_keys(self.bookmark_title_changed)
+        button = self.browser.find_element(By.ID, value="button-save")
         button.click()
-        time.sleep(1)
+        time.sleep(2)
+
+        table = self.browser.find_element(By.ID, value="table-bookmarks")
+        self.assertIn(self.bookmark_title_changed, table.text)
+
+    def do_test_download_bookmark(self):
+        """
+        Test video download for added bookmark.
+        - Click "Bookmarks" in the navbar.
+        - Open "Edit bookmark" modal for added URL, click "Download", close modal.
+        - Wait for download to end, open modal, check if downloaded file displayed.
+        """
+        self.click_navbar_menu_bookmarks()
+        self.click_bookmarks_table_first_row_edit_button()
+
+        # Click "Download"
+        button = self.browser.find_element(By.ID, value="button-download")
+        button.click()
+
+        # Click "Close" and wait for the download
+        button = self.browser.find_element(By.ID, value="button-close")
+        button.click()
+        time.sleep(15)
+
+        # Click "Refresh"
+        button = self.browser.find_element(By.ID, value="button-refresh")
+        button.click()
+        time.sleep(3)
+
+        self.click_bookmarks_table_first_row_edit_button()
+
+        # Check if file info displayed
+        span = self.browser.find_element(By.ID, value="downloaded-file-info")
+        self.assertIn("Video file", span.text)
+
+        # Click "Close"
+        button = self.browser.find_element(By.ID, value="button-close")
+        button.click()
+
+    def do_test_delete_bookmark(self):
+        self.click_navbar_menu_bookmarks()
+        self.click_bookmarks_table_first_row_edit_button()
 
         # Click "Delete" button, confirmation modal will popup
         button = self.browser.find_element(By.ID, value="button-delete")
         button.click()
+        time.sleep(1)
 
         # Wait for modal to show, and "click" OK
         WebDriverWait(self.browser, 10).until(EC.alert_is_present())
         self.browser.switch_to.alert.accept()
+        time.sleep(1)
 
-        self.assertNotIn(self.bookmark_title, self.browser.title)
+        # Check that deleted bookmark is not listed on the page (there must be no table at all)
+        table = self.browser.find_element(By.TAG_NAME, value="body")
+        self.assertNotIn(self.bookmark_title, table.text)
 
-        # self.fail("End the test.")
-        time.sleep(3)
+        time.sleep(1)